vboot: When hashing the RW firmware, handle 64KB chunks at a time.
It's a little simpler to load the whole RW firmware in one go and then pass it
all through to get hashed, but that means that we need to allocate a large
buffer to hold the RW image. If the heap is expanded out (like it is for UEFI,
for instance) then the RW image will necessarily be bigger than the heap it
contains, and allocating a buffer for the RW image will necessarily fail,
unless the RO heap can be made big enough independently.

It's easiest to just pass the RW firmware to the hasher in smaller chunks and
should have minimal impact on performance. It might even actually perform
better because of improved cache locality.

Change-Id: I14eb8c6017ada3ddb9369274cbc8288444dce257