UEFI: Seperate out the bss.
gnu-efi claimed that EFI "doesn't seem to like a .bss section", but I don't
think that's true. With the .bss at the end of the image, the PE+ image will
not allocate space for the bss, but EFI will set aside memory for it when it's
loaded.

I checked that the space is actually set aside and not just accessible by
printing the memory map returned by EFI boot services, and also checking the
"loaded image protocol" attached to depthcharge's image handle to see what the
image base and size were according to EFI.

There are two benefits for not having space allocated for the bss in the image.
First, it means there's less data to load from disk and for vboot to hash.
Second, it means that the RO and RW images are no longer going to be strictly
larger than their heaps (since their heaps are part of the bss) and so buffers
can be malloc-ed which can hold them.

One downside is that the RW firmware images will no longer be XIP, since the
.bss won't be contained in the image. Additional space will need to be
allocated, and the image will need to be loaded into it before it's run.

Change-Id: I1c391c35483e8ae382b320e604e01e6a5d94cae8