module: Abstract the module starting process.
Most modules will start as compressed images loaded to a particular address
by a trampoline with no external entity managing memory. That's not true on
UEFI systems, though, and may not be true on other systems in the future. The
cleanest way to abstract that behavior seems to be by making a driver like
interface for starting a module, and using that instead of starting a module
directly in generic code. Code which is specific to a particular device or for
use in a scenario where we can assume the generic module code will work can
still use the raw start_module function, particularly if it can't allocate a
callback structure on the heap as new_dc_module() does.

This might also be useful in the future when there are modules which need to
be verified further before they're started in a device specific way, for
instance if the hash of a module which implements developer mode has been
verified and needs to be used to verify the actual developer mode module
before it runs. That detail can be abstracted away by the module starting
callback and handled transparently.

Change-Id: Ia58d53072bcac350815fab11c250b1b3572ca487