module: Wrap the call to the lzma utility in a python script.
This script will fix up the resulting compressed file so that it has the
uncompressed size in its header. The original lzma utility would put that size
in there if the source had a well defined size up front, for instance a regular
file. On some systems (notably the ChromeOS chroot, probably inherited from
Gentoo) the uncompressed size is always left as "unknown", regardless of what
the source file was. This makes it hard to decompress the file in depthcharge
since we can't tell how big a buffer we'll need for the decompressed data.

Change-Id: If80723c82f2c8ae587d53f36a5300b46b464f2b1