lzma: Add a function to extract the expanded size of an lzma stream.
This information is in the LZMA header. It's useful when allocating a buffer
to hold the decompressed data. This hasn't been an issue before because we had
pretty much all of memory to play with, but when running in an environment like
UEFI, we can't assume we'll have essentially unbounded regions of memory to
drop the decompressed data in.

Change-Id: I4d34856efeb9a78cf11a1426394396a69fac4050