UEFI: blockdev: Add a UEFI block IO wrapper driver.
A given controller has to be hooked up as either fixed or removable, but the
UEFI block IO protocol is applied to both. It would be nice to tidy up the
blockdev interface to address this and some other minor problems, but in the
mean time, a given UEFI blockdev controller object will filter out media to be
either fixed or removable, depending on its flavor.

Change-Id: I7989fb873312aec44bccc49220dfc5a54e2eb44b