LZMA: Handle unknown decompressed size in the decompression function.
If you're explicitly trying to get the decompressed size then not having one
should be considered an error. If you're just trying to decompress some data,
as long as you're not going to overflow into something else, it should be ok
to not know how big the result will be and to just stop when you run out of
room.

Change-Id: I35c5916c3d8afd2e85b2432a1453a757351d84fb