x86: Copy the command line and boot params into place after handoff cleanups.
When running under UEFI, the handoff cleanup function will be where
ExitBootservices is called, and so is the boundary between where we have to
respect UEFI's boot time boundaries and when we're in full control of the
machine. We want it to be as late as possible to still be able to use UEFI's
drivers.

On systems which aren't hosted by some other firmware or an OS, this
constraint is still simple and pretty easy to follow, and shouldn't impact
performance at all.

Change-Id: I4facfa3ebf545c80875131579eb9c5265216ecef