uefi: Make installing the exit boot services cleanup automatic.
If the UEFI system table is successfully retrieved, then there's a chance
the exit boot services cleanup is necessary, and if not then it isn't. Since
it's really important the cleanup is installed, but only when actually
necessary, it's better for it to be handled automatically than to clutter
individual implementations with it, or give them a chance to forget it.

Change-Id: Ic1022befd3c421f8b2b38435aa432c9f59a67602